+++ hpcaccess/auth_backends.py
@@ -13,15 +13,10 @@ LDAP2_DOMAIN = getattr(settings, "AUTH_LDAP2_USERNAME_DOMAIN", None)
 class PrimaryLDAPBackend(LDAPBackend):
     settings_prefix = "AUTH_LDAP_"
 
-    def authenticate(
-        self, request=None, username=None, password=None, **kwargs
-    ):
+    def authenticate(self, request=None, username=None, password=None, **kwargs):
         # Login with username@DOMAIN
         if LDAP_DOMAIN:
-            if (
-                username.find("@") == -1
-                or username.strip().split("@")[1].upper() != LDAP_DOMAIN
-            ):
+            if username.find("@") == -1 or username.strip().split("@")[1].upper() != LDAP_DOMAIN:
                 return None
             ldap_user = _LDAPUser(self, username=username.split("@")[0].strip())
         # Login with username only
@@ -45,13 +40,8 @@ class PrimaryLDAPBackend(LDAPBackend):
 class SecondaryLDAPBackend(LDAPBackend):
     settings_prefix = "AUTH_LDAP2_"
 
-    def authenticate(
-        self, request=None, username=None, password=None, **kwargs
-    ):
-        if (
-            username.find("@") == -1
-            or username.split("@")[1].upper() != LDAP2_DOMAIN
-        ):
+    def authenticate(self, request=None, username=None, password=None, **kwargs):
+        if username.find("@") == -1 or username.split("@")[1].upper() != LDAP2_DOMAIN:
             return None
 
         ldap_user = _LDAPUser(self, username=username.split("@")[0].strip())