+++ usersec/tests/test_forms.py
@@ -20,9 +20,7 @@ class TestHpcGroupCreateRequestForm(TestCase):
         data_invalid = {**self.data_valid, "resources_requested": {}}
         form = HpcGroupCreateRequestForm(data=data_invalid)
         self.assertFalse(form.is_valid())
-        self.assertEqual(
-            form.errors["resources_requested"], ["This field is required."]
-        )
+        self.assertEqual(form.errors["resources_requested"], ["This field is required."])
 
     def test_form_invalid_resources_requested_not_dict(self):
         data_invalid = {
@@ -46,17 +44,13 @@ class TestHpcGroupCreateRequestForm(TestCase):
         data_invalid = {**self.data_valid, "description": {}}
         form = HpcGroupCreateRequestForm(data=data_invalid)
         self.assertFalse(form.is_valid())
-        self.assertEqual(
-            form.errors["description"], ["This field is required."]
-        )
+        self.assertEqual(form.errors["description"], ["This field is required."])
 
     def test_form_valid_hpcadmin(self):
         user_hpcadmin = self.make_user("hpcadmin")
         user_hpcadmin.is_hpcadmin = True
         user_hpcadmin.save()
-        form = HpcGroupCreateRequestForm(
-            user=user_hpcadmin, data=self.data_valid
-        )
+        form = HpcGroupCreateRequestForm(user=user_hpcadmin, data=self.data_valid)
         self.assertTrue(form.is_valid())
 
     def test_form_invalid_hpcadmin_comment_missing(self):